backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
alembic
pydantic>=2.7,<3
pydantic-settings
python-dotenv
httpx
groq
openai
google-genai
duckduckgo-search
python-multipart
python-docx
reportlab
Pillow
aiofiles